+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Use Bract 0.5.1
   - [TODO] Use the GA version
 - Add fallback config (in `baseconfig.edn`) for application version and hostname
+- Use `bract.cli.main` as CLI entry point
 - Use soft-delete to remove TODO items in database
 - [TODO] Add CLI command to purge soft-deleted TODO items